# src/Dle.Control/openapi/dle-control.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Deep Link Engine — control plane",
    "description": "Link, domain, application, tenant and credential management, plus attribution and analytics reporting. Errors are RFC 9457 problem documents whose type member is one of the stable identifiers documented at https://docs.dle.dev/problems/. Writes accept an Idempotency-Key header: the same key with the same body replays the stored response, and the same key with a different body is a conflict.",
    "version": "v1"
  },
  "paths": {
...
    "/v1/resolve": {
      "post": {
        "tags": [
          "SDK"
        ],
        "summary": "Recovers the click context of a fresh installation.",
        "description": "Called once on the first launch after an install. Answers with the strategy that produced the match and how much to trust it; an unmatched install is a 200 with matched=false, not an error. Limited to five calls an hour per install_id (§E.9).",
        "operationId": "ResolveInstall",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/events": {
      "post": {
        "tags": [
          "SDK"
        ],
        "summary": "Records a batch of client events, link opens included.",
        "description": "Batched, at most 100 events per request. The link_open event is what makes a direct open through a verified Universal or App Link measurable at all: the operating system opens the application without any request reaching the engine, so without this report the most successful campaigns are the most under-counted (§B.6.4, FR-223). A link_open whose URL resolves to one of the caller's links creates a direct_open attribution with confidence 1.00.",
        "operationId": "RecordSdkEvents",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventBatchDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventBatchAcceptedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/claim-codes": {
      "post": {
        "tags": [
          "SDK"
        ],
        "summary": "Issues the six characters an interstitial page shows the user.",
        "description": "The deterministic iOS path of ADR-008. The edge has just written a click event and exchanges its identifier for a short lived, single use code the user can read off the screen and type into the application (FR-184, §B.6.3). Only the keyed hash of the code is stored; the plaintext exists in this response and nowhere else.",
        "operationId": "IssueClaimCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueClaimCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimCodeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResolveRequestDto": {
        "required": [
          "install_id",
          "platform"
        ],
        "type": "object",
        "properties": {
          "install_id": {
            "type": "string",
            "description": "Identifier the SDK generated and persisted for this installation. Deduplication and\r\n            replay protection key off it (FR-188, TC-143)."
          },
          "platform": {
            "type": "string",
            "description": "Client platform: `ios`, `android`, `desktop` or `other`."
          },
          "app_version": {
            "type": [
              "null",
              "string"
            ],
            "description": "Host application version, for version scoped routing rules (FR-124)."
          },
          "os_version": {
            "type": [
              "null",
              "string"
            ],
            "description": "Operating system version reported by the device."
          },
          "referrer": {
            "type": [
              "null",
              "string"
            ],
            "description": "Raw Google Play Install Referrer string. The deterministic Android path (S1, FR-182)."
          },
          "claim_code": {
            "type": [
              "null",
              "string"
            ],
            "description": "Code the user typed from the interstitial page. A deterministic iOS path (S3, FR-184)."
          },
          "login_key": {
            "type": [
              "null",
              "string"
            ],
            "description": "Opaque, already hashed account identifier used to reconcile a web session with a\r\n            signed in user. A deterministic path on both platforms (S2, FR-185)."
          },
          "signals": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Device signals for probabilistic matching. Sent only when the module is enabled and\r\n            attribution consent was given; otherwise the SDK omits the object entirely (TC-145, TC-146).",
                "$ref": "#/components/schemas/DeviceSignalsDto"
              }
            ]
          },
          "consent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "Recorded consent. Absent or negative consent disables every non-deterministic\r\n            strategy and suppresses storage of the signals (ePrivacy art. 5(3), §E.6.2).",
                "$ref": "#/components/schemas/ConsentDto"
              }
            ]
          }
        },
        "description": "Body of `POST /v1/resolve`, the call an SDK makes once on the first launch after install\r\nto recover the context of the click that led to it (§B.7.2, UC-04, UC-05)."
      },
      "ResolveResponseDto": {
        "required": [
          "matched",
          "match_type",
          "confidence"
        ],
        "type": "object",
        "properties": {
          "matched": {
            "type": "boolean",
            "description": "Whether any strategy matched. When false the app continues its normal onboarding."
          },
          "match_type": {
            "type": "string",
            "description": "Strategy that produced the match: `none`, `install_referrer`, `login`,\r\n            `claim_code`, `probabilistic` or `direct_open`."
          },
          "confidence": {
            "type": "number",
            "description": "Confidence from 0.00 to 1.00. Exactly 1.00 for deterministic strategies.",
            "format": "double"
          },
          "click_id": {
            "type": [
              "null",
              "string"
            ],
            "description": "Identifier of the matched click."
          },
          "link": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "description": "The matched link, when there is one.",
                "$ref": "#/components/schemas/ResolveLinkDto"
              }
            ]
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Parameters handed to the application, typically the UTM set of the link plus custom data."
          },
          "expires_in": {
            "type": "integer",
            "description": "Seconds this result stays valid. Zero means it is final and must not be re-requested.",
            "format": "int32"
          }
        },
        "description": "Response to `POST /v1/resolve`. Always states how the match was made and how much to trust\r\nit: a probabilistic result is never dressed up as a certainty (FR-186, ADR-008)."
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "description": "A control-plane API key, presented as a bearer token or in the X-Api-Key header. The key looks like dle_<prefix>_<secret>; only its Argon2id hash is stored, so a lost key is replaced rather than recovered.",
        "scheme": "bearer"
      },
      "SdkKey": {
        "type": "apiKey",
        "description": "A key embedded in a customer application. It authenticates the two SDK ingestion routes and nothing else, and must be assumed extractable from the application binary.",
        "name": "X-Dle-Sdk-Key",
        "in": "header"
      }
    }
  },
...
}